Вынес вспомогательные функции в untils
