+++ src/view/popup.js
@@ -1,3 +1,5 @@
+import {convertsDate} from "../utils.js";
+
 const createGenresTemplate = (genre) => {
   const genreTemplate = genre.map((element) => {
     return `<span class="film-details__genre">${element}</span>`;
@@ -12,16 +14,6 @@ const createGenreContainerTemplate = (genre) => {
       ${createGenresTemplate(genre)}
     </td>`
   );
-};
-
-const convertsDate = (date) => {
-  const day = String(date.getDate()).padStart(2, `0`);
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, `0`);
-  const minutes = String(date.getMinutes()).padStart(2, `0`);
-  const second = String(date.getSeconds()).padStart(2, `0`);
-  const convertedDate = `${year}/${month}/${day} ${minutes}:${second}`;
-  return convertedDate;
 };
 
 const createCommentsTemplate = (comments) => {
